third_party/blink/web_tests/virtual/threaded/external/wpt/css/css-scroll-snap/snap-after-relayout/resnap-to-focused-expected.txt

This is a testharness.js-based test.
FAIL Resnap to focused element after relayout assert_equals: After resize, should snap to row 4. expected 4 but got 3
Harness: the test ran to completion.
